% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastStudentT.R
\name{fastStudentT}
\alias{fastStudentT}
\title{Estimate parameters of Student t distribution}
\usage{
fastStudentT(y, df = 50, tol = 0.001, verbose = TRUE)
}
\arguments{
\item{y}{vector of observed data}

\item{df}{degrees of freedom used to initialize the algorithm}

\item{tol}{convergence tolerance for the log likelihood}

\item{verbose}{show messages}
}
\description{
Estimation is substantially faster for large datasets than other methods
}
